// src/lib/region/epimap/types.ts
export interface Epimap {
  chromosome: string;
  start_position: number;
  end_position: number;
  category: string;
  bssid: string;
  newgroup: string;
  extended_info: string;
  lifestage: string;
  sex: string;
  type: string;
  sample_name: string;
  color: string;
  state_full_name: string;
}